refactor(gui): moved chipselect to forms directory and updated consumer imports
