+++ frontend/src/js/common-ui/forms/ChipSelect.test.tsx
@@ -11,20 +11,13 @@
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
-import { FormProvider, useForm } from 'react-hook-form';
-
-import { render } from '@/testUtils';
+import { formRenderWrapper } from '@/testUtils';
 import { undefineds } from '@northern.tech/testing/mockData';
 import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { vi } from 'vitest';
 
 import ChipSelect from './ChipSelect';
-
-const FormWrapper = ({ children }) => {
-  const methods = useForm({ mode: 'onChange', defaultValues: { tags: [] } });
-  return <FormProvider {...methods}>{children}</FormProvider>;
-};
 
 const defaultProps = {
   name: 'tags',
@@ -33,35 +26,25 @@ const defaultProps = {
   placeholder: 'Select tags'
 };
 
+const formConfig = { mode: 'onChange', defaultValues: { tags: [] } };
+
 describe('ChipSelect Component', () => {
   it('renders correctly', async () => {
-    const { baseElement } = render(
-      <FormWrapper>
-        <ChipSelect {...defaultProps} />
-      </FormWrapper>
-    );
+    const { baseElement } = formRenderWrapper(<ChipSelect {...defaultProps} />, formConfig);
     const view = baseElement.firstChild.firstChild;
     expect(view).toMatchSnapshot();
     expect(view).toEqual(expect.not.stringMatching(undefineds));
   });
 
   it('renders with disabled state as readOnly', async () => {
-    render(
-      <FormWrapper>
-        <ChipSelect {...defaultProps} disabled />
-      </FormWrapper>
-    );
+    formRenderWrapper(<ChipSelect {...defaultProps} disabled />, formConfig);
     const input = screen.getByRole('combobox');
     expect(input).toHaveAttribute('readonly');
   });
 
   it('allows selecting an option from the dropdown', async () => {
     const user = userEvent.setup({ advanceTimers: vi.advanceTimersByTime });
-    render(
-      <FormWrapper>
-        <ChipSelect {...defaultProps} />
-      </FormWrapper>
-    );
+    formRenderWrapper(<ChipSelect {...defaultProps} />, formConfig);
     const input = screen.getByRole('combobox');
     await user.click(input);
     await user.click(screen.getByText('alpha'));
@@ -70,11 +53,7 @@ describe('ChipSelect Component', () => {
 
   it('adds comma-separated values as chips', async () => {
     const user = userEvent.setup({ advanceTimers: vi.advanceTimersByTime });
-    render(
-      <FormWrapper>
-        <ChipSelect {...defaultProps} />
-      </FormWrapper>
-    );
+    formRenderWrapper(<ChipSelect {...defaultProps} />, formConfig);
     const input = screen.getByRole('combobox');
     await user.type(input, 'foo,bar,');
     await waitFor(() => expect(screen.getByText('foo')).toBeInTheDocument());
@@ -83,11 +62,7 @@ describe('ChipSelect Component', () => {
 
   it('adds whitespace-separated values as chips on blur', async () => {
     const user = userEvent.setup({ advanceTimers: vi.advanceTimersByTime });
-    render(
-      <FormWrapper>
-        <ChipSelect {...defaultProps} />
-      </FormWrapper>
-    );
+    formRenderWrapper(<ChipSelect {...defaultProps} />, formConfig);
     const input = screen.getByRole('combobox');
     await user.type(input, 'one two');
     await user.tab();
