+++ frontend/src/js/components/deployments/deployment-wizard/ReleaseArtifactFilter.tsx
@@ -19,11 +19,11 @@ import { Close as CloseIcon, ExpandLess as ExpandLessIcon, ExpandMore as ExpandM
 import { Button, DialogContent, Divider, Typography, menuItemClasses } from '@mui/material';
 import { makeStyles } from 'tss-react/mui';
 
-import ChipSelect from '@northern.tech/common-ui/ChipSelect';
 import EnterpriseNotification from '@northern.tech/common-ui/EnterpriseNotification';
 import { Link } from '@northern.tech/common-ui/Link';
 import { ControlledSearch } from '@northern.tech/common-ui/Search';
 import { BaseDialog } from '@northern.tech/common-ui/dialogs/BaseDialog';
+import ChipSelect from '@northern.tech/common-ui/forms/ChipSelect';
 import { ControlledSelect } from '@northern.tech/common-ui/forms/ControlledSelect';
 import { BENEFITS } from '@northern.tech/store/constants';
 import type { SoftwareKind } from '@northern.tech/store/releasesSlice';
